requirements.txt
# requirements.txt
streamlit
tensorflow
numpy
opencv_python
Pillow